streamlit-ui/requirements.txt
# Core Framework
streamlit>=1.28.0

# Deep Learning
tensorflow>=2.10.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Image Processing
Pillow>=9.0.0

# Data Processing
pandas>=1.3.0

# Database
pymongo>=4.0.0

# Authentication & Security
bcrypt>=3.2.0
email-validator>=1.1.0

# Environment & Configuration
python-dotenv>=0.19.0

# Data Visualization
plotly>=5.3.0

# Web & File Handling
python-multipart>=0.0.5
requests>=2.25.0

# Utilities
python-dateutil>=2.8.0
pytz>=2021.1